fixing update

+++ controller.go
@@ -10,13 +10,16 @@ func CreateTodo(c *fiber.Ctx) error {
 		return err
 	}
 	db.Create(&todos)
-	return c.JSON(todos)
+	return c.JSON(fiber.Map{"messages": "Todo created", "data": todos})
 }
 
 func GetTodo(c *fiber.Ctx) error {
 	var todos []Todo
 	db.Find(&todos)
-	return c.JSON(todos)
+	return c.JSON(fiber.Map{
+		"data":    todos,
+		"message": "Todos fetched",
+	})
 }
 
 func UpdateTodo(c *fiber.Ctx) error {
@@ -34,9 +37,14 @@ func UpdateTodo(c *fiber.Ctx) error {
 	todos.Task = UpdateTodo.Task
 	todos.Completed = UpdateTodo.Completed
 
+	if todos.Id == 0 {
+		return c.Status(404).JSON(fiber.Map{
+			"message": "Todo not found",
+		})
+	}
 	db.Save(&todos)
 
-	return c.JSON(todos)
+	return c.JSON(fiber.Map{"messages": "Todo updated", "data": todos})
 }
 
 func DeleteTodo(c *fiber.Ctx) error {
@@ -46,7 +54,13 @@ func DeleteTodo(c *fiber.Ctx) error {
 
 	db.Find(&todos, "id = ?", id)
 
+	if todos.Id == 0 {
+		return c.Status(404).JSON(fiber.Map{
+			"message": "Todo not found",
+		})
+	}
+
 	db.Delete(&todos, "id = ?", id)
 
-	return c.JSON("Todo deleted")
+	return c.JSON(fiber.Map{"messages": "Todo deleted"})
 }
